Move kafka metrics to separate instrumentation module
